requirements.txt
plotly<=5.24.1
dash
pandas
PyYAML
dash_bootstrap_components==1.7.1
mrcfile
h5py
dash-slicer
napari[pyqt5]
starfile
